Fixed join with advice segfault.

+++ path.cc
@@ -166,13 +166,13 @@ bool Path::isPartlyDisjoint(const Path &p, int dist) const {
   unordered_set<int> nodes_ids;
 
   int len = 0;
-  for (size_t i = 0; i < nodes_.size() && len < dist; i++) {
+  for (int i = 0; i < nodes_.size() && len < dist; i++) {
     const auto &node = nodes_[i];
     nodes_ids.insert(node->id_);
     len += node->str_.size();
   }
   len = 0;
-  for (size_t i = nodes_.size() - 1; i >= 0 && len < dist; i--) {
+  for (int i = (int)nodes_.size() - 1; i >= 0 && len < dist; i--) {
     const auto &node = nodes_[i];
     nodes_ids.insert(node->id_);
     len += node->str_.size();
@@ -180,13 +180,13 @@ bool Path::isPartlyDisjoint(const Path &p, int dist) const {
 
   // checking begin and end of path p
   len = 0;
-  for (size_t i = 0; i < p.nodes_.size() && len < dist; i++) {
+  for (int i = 0; i < p.nodes_.size() && len < dist; i++) {
     const auto &node = p.nodes_[i];
     if (nodes_ids.count(node->id_) > 0 || nodes_ids.count(node->rc_->id_) > 0) return false;
     len += node->str_.size();
   }
   len = 0;
-  for (size_t i = p.nodes_.size() - 1; i >= 0 && len < dist; i--) {
+  for (int i = (int)p.nodes_.size() - 1; i >= 0 && len < dist; i--) {
     const auto &node = p.nodes_[i];
     if (nodes_ids.count(node->id_) > 0 || nodes_ids.count(node->rc_->id_) > 0) return false;
     len += node->str_.size();
